# docs/phase-4-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "London Biodiversity Expedition Planner API",
    "description": "A privacy-bounded API for London bird-expedition evidence, agent execution, human decisions, and checkpoint time travel.",
    "version": "4.0.0"
  },
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_v1_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "type": "array",
                  "title": "Response List Runs Api V1 Runs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{thread_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__thread_id__get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{thread_id}/history": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get History",
        "operationId": "get_history_api_v1_runs__thread_id__history_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{thread_id}/checkpoints/{checkpoint_id}/state": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Checkpoint State",
        "operationId": "get_checkpoint_state_api_v1_runs__thread_id__checkpoints__checkpoint_id__state_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "checkpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkpoint Id"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Node Id"
            }
          },
          {
            "name": "graph_step",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Graph Step"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{thread_id}/checkpoints/{checkpoint_id}/evidence": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Checkpoint Evidence",
        "operationId": "get_checkpoint_evidence_api_v1_runs__thread_id__checkpoints__checkpoint_id__evidence_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "checkpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkpoint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{thread_id}/checkpoints/{checkpoint_id}/map": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Checkpoint Map",
        "operationId": "get_checkpoint_map_api_v1_runs__thread_id__checkpoints__checkpoint_id__map_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "checkpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkpoint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapEvidenceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{thread_id}/resume": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Resume Run",
        "operationId": "resume_run_api_v1_runs__thread_id__resume_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{thread_id}/replay": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Replay Run",
        "operationId": "replay_run_api_v1_runs__thread_id__replay_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{thread_id}/fork": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Fork Run",
        "operationId": "fork_run_api_v1_runs__thread_id__fork_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForkRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{thread_id}/compare": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Compare Runs",
        "operationId": "compare_runs_api_v1_runs__thread_id__compare_get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "checkpoint_a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Checkpoint A"
            }
          },
          {
            "name": "checkpoint_b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Checkpoint B"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComparisonView": {
        "properties": {
          "comparison": {
            "$ref": "#/components/schemas/PlanComparison"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "comparison"
        ],
        "title": "ComparisonView"
      },
...
      "CreateRunRequest": {
        "properties": {
          "request": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Request"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 96,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "fixture",
              "live"
            ],
            "title": "Data Mode",
            "default": "fixture"
          },
          "model_mode": {
            "type": "string",
            "enum": [
              "scripted",
              "live"
            ],
            "title": "Model Mode",
            "default": "scripted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request"
        ],
        "title": "CreateRunRequest"
      },
...
      "ForkRunRequest": {
        "properties": {
          "checkpoint_id": {
            "type": "string",
            "minLength": 1,
            "title": "Checkpoint Id"
          },
          "updates": {
            "$ref": "#/components/schemas/ForkUpdates"
          },
          "branch_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "checkpoint_id",
          "updates"
        ],
        "title": "ForkRunRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryView": {
        "properties": {
          "checkpoints": {
            "items": {
              "$ref": "#/components/schemas/CheckpointSummary"
            },
            "type": "array",
            "title": "Checkpoints"
          },
          "branches": {
            "items": {
              "$ref": "#/components/schemas/RunBranch"
            },
            "type": "array",
            "title": "Branches"
          },
          "executions": {
            "items": {
              "$ref": "#/components/schemas/RunExecution"
            },
            "type": "array",
            "title": "Executions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "checkpoints",
          "branches",
          "executions"
        ],
        "title": "HistoryView"
      },
...
      "OperationAccepted": {
        "properties": {
          "operation_id": {
            "type": "string",
            "minLength": 1,
            "title": "Operation Id"
          },
          "thread_id": {
            "type": "string",
            "maxLength": 96,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Thread Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status",
            "default": "queued"
          },
          "events_url": {
            "type": "string",
            "minLength": 1,
            "title": "Events Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "thread_id",
          "events_url"
        ],
        "title": "OperationAccepted"
      },
...
      "ReplayRunRequest": {
        "properties": {
          "checkpoint_id": {
            "type": "string",
            "minLength": 1,
            "title": "Checkpoint Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "checkpoint_id"
        ],
        "title": "ReplayRunRequest"
      },
...
      "RunDetail": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunSummary"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/OperationView"
            },
            "type": "array",
            "title": "Operations"
          },
          "branches": {
            "items": {
              "$ref": "#/components/schemas/RunBranch"
            },
            "type": "array",
            "title": "Branches"
          },
          "executions": {
            "items": {
              "$ref": "#/components/schemas/RunExecution"
            },
            "type": "array",
            "title": "Executions"
          },
          "pending_decisions": {
            "items": {
              "$ref": "#/components/schemas/PendingDecisionView"
            },
            "type": "array",
            "title": "Pending Decisions"
          },
          "final_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinalPlanView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run"
        ],
        "title": "RunDetail"
      },
...
      "RunSummary": {
        "properties": {
          "thread_id": {
            "type": "string",
            "maxLength": 96,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Thread Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "waiting_for_input",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "data_mode": {
            "type": "string",
            "enum": [
              "fixture",
              "live"
            ],
            "title": "Data Mode"
          },
          "model_mode": {
            "type": "string",
            "enum": [
              "scripted",
              "live"
            ],
            "title": "Model Mode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "current_checkpoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Checkpoint Id"
          },
          "branch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "execution_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Id"
          },
          "interrupt_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt Kind"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "thread_id",
          "status",
          "data_mode",
          "model_mode",
          "created_at",
          "updated_at"
        ],
        "title": "RunSummary"
      },
...
      "StateView": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "thread_id": {
            "type": "string",
            "minLength": 1,
            "title": "Thread Id"
          },
          "branch_id": {
            "type": "string",
            "minLength": 1,
            "title": "Branch Id"
          },
          "execution_id": {
            "type": "string",
            "minLength": 1,
            "title": "Execution Id"
          },
          "checkpoint_id": {
            "type": "string",
            "minLength": 1,
            "title": "Checkpoint Id"
          },
          "parent_checkpoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Checkpoint Id"
          },
          "node_id": {
            "type": "string",
            "minLength": 1,
            "title": "Node Id"
          },
          "graph_step": {
            "type": "integer",
            "title": "Graph Step"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "next_nodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Nodes"
          },
          "terminal_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Status"
          },
          "request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StateRequestView"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StateLocationView"
              },
              {
                "type": "null"
              }
            ]
          },
          "taxon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StateTaxonView"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "$ref": "#/components/schemas/StateEvidenceView"
          },
          "plan": {
            "$ref": "#/components/schemas/StatePlanView"
          },
          "hitl": {
            "$ref": "#/components/schemas/StateHitlView"
          },
          "counters": {
            "$ref": "#/components/schemas/StateCountersView"
          },
          "invalidated_evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Invalidated Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "thread_id",
          "branch_id",
          "execution_id",
          "checkpoint_id",
          "node_id",
          "graph_step",
          "created_at",
          "evidence",
          "plan",
          "hitl",
          "counters"
        ],
        "title": "StateView",
        "description": "Strictly allow-listed, coordinate-free view of one saved checkpoint."
      },
...
    }
  }
}